# lib/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Api
  version: 1.0.0
  description: |
    Typed contract for the Complete iQ Carrier Audit API.

    Authenticated requests use the `sid` HTTP-only session cookie. The optional
    `X-Organization-Id` header selects one of the caller's organization
    memberships; otherwise the default (or earliest) membership is used.
    Organization-scoped lookups deliberately return `404` when a resource
    belongs to another tenant so that cross-tenant identifiers are not
    disclosed.

    Job-enqueueing operations that accept `X-Idempotency-Key` scope that value
    to the selected organization and operation inputs. Repeating an equivalent
    request returns the existing durable processing job and sets `duplicate`
    to `true` when that field is present.
servers:
  - url: /api
    description: Application API
security:
  - cookieAuth: []
...
paths:
...
  /ingest:
    post:
      operationId: ingestClaimFile
      tags: [Intake, Jobs]
      summary: Upload and enqueue a claim file
      description: |
        Stores a file, creates a processing claim and document, and enqueues an
        `ingest` job. Requires `claims:create`. The idempotency identity includes
        the selected organization, file SHA-256, requested carrier, and optional
        caller key. Equivalent requests return the existing job with
        `duplicate: true`.
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/IngestRequest"
      responses:
        "202":
          description: File accepted and durable processing job created or reused.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /claims/{id}/processing-status:
    get:
      operationId: getClaimProcessingStatus
      tags: [Intake, Jobs]
      summary: Get the latest claim processing status
      description: |
        Requires `jobs:read`. Maps the latest durable job to the frontend
        `processing`, `error`, or `ready` state.
      parameters:
        - $ref: "#/components/parameters/ClaimId"
        - $ref: "#/components/parameters/OrganizationId"
      responses:
        "200":
          description: Current claim workflow and latest job state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessingStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /claims/{id}/retry:
    post:
      operationId: retryClaimProcessing
      tags: [Intake, Jobs]
      summary: Retry claim processing
      description: |
        Requires `jobs:retry`. A latest failed, degraded, or cancelled job is
        reset in place; otherwise an equivalent retry job is created or reused.
        `X-Idempotency-Key` participates when a new retry job is needed.
      parameters:
        - $ref: "#/components/parameters/ClaimId"
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: Retry queued or an active equivalent job returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnqueueJobResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /claims/{id}/reprocess:
    post:
      operationId: reprocessClaim
      tags: [Intake, Jobs]
      summary: Reprocess a claim with a carrier
      description: |
        Requires `audits:run`. Enqueues or reuses a `reprocess` job for the latest
        claim-file document. The idempotency identity includes the carrier and
        either `X-Idempotency-Key` or the current successful audit identity.
      parameters:
        - $ref: "#/components/parameters/ClaimId"
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReprocessClaimRequest"
      responses:
        "202":
          description: Reprocessing queued or an equivalent active job returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnqueueJobResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sid
      description: HTTP-only server session cookie, marked Secure outside development.
    inboundToken:
      type: apiKey
      in: query
      name: token
      description: Shared secret configured for the SendGrid Inbound Parse callback.
  parameters:
    OrganizationId:
      name: X-Organization-Id
      in: header
      required: false
      description: |
        Organization membership to use for this request. Omit to use the
        caller's default membership. An inaccessible value returns `403`.
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: X-Idempotency-Key
      in: header
      required: false
      description: |
        Caller-provided idempotency key. It is combined with organization and
        operation-specific inputs; equivalent requests reuse the durable job.
      schema:
        type: string
        minLength: 1
        maxLength: 255
    ClaimId:
      name: id
      in: path
      required: true
      description: Claim identifier.
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Request parameters or body failed route validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: A valid session or organization membership is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The caller lacks the selected organization or required permission.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: |
        Resource not found in the selected organization. Organization-scoped
        routes use this response for cross-tenant identifiers.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    PayloadTooLarge:
      description: Uploaded content exceeds the route limit.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequests:
      description: Request rate limit exceeded.
      headers:
        RateLimit-Limit:
          schema:
            type: integer
        RateLimit-Remaining:
          schema:
            type: integer
        RateLimit-Reset:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerError:
      description: The server could not complete the request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      description: Shared JSON error payload.
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
        code:
          type: string
          description: Optional stable machine-readable code.
        details:
          description: Optional structured diagnostic context.
          nullable: true
        requestId:
          type: string
          description: Optional request correlation identifier.
      additionalProperties: false
...
    ClaimStatus:
      type: string
      enum: [processing, pending, analyzed, approved, denied, error, archived]
    SystemWorkflowStatus:
      type: string
      enum: [uploaded, processing, ready, error, archived]
    AiWorkflowStatus:
      type: string
      enum: [not_started, queued, running, succeeded, degraded, failed, cancelled]
    HumanReviewStatus:
      type: string
      enum: [unassigned, pending, in_review, approved, changes_requested]
...
    ProcessingJobType:
      type: string
      enum: [ingest, audit, retry, reprocess, extract]
    ProcessingJobStatus:
      type: string
      enum: [queued, running, succeeded, degraded, failed, cancelled]
    ProcessingJobStage:
      type: string
      enum: [uploaded, scanning, extracting, auditing, degraded, completed, failed, cancelled]
...
    JobError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
      additionalProperties: false
    ProcessingJobSummary:
      type: object
      required: [id, status, stage]
      properties:
        id:
          type: string
          format: uuid
        claimId:
          type: string
          format: uuid
          nullable: true
        documentId:
          type: string
          format: uuid
          nullable: true
        type:
          $ref: "#/components/schemas/ProcessingJobType"
        status:
          $ref: "#/components/schemas/ProcessingJobStatus"
        stage:
          $ref: "#/components/schemas/ProcessingJobStage"
        progress:
          type: integer
          minimum: 0
          maximum: 100
        attemptCount:
          type: integer
          minimum: 0
        maxAttempts:
          type: integer
          minimum: 1
        error:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/JobError"
      additionalProperties: false
...
    EnqueueJobResponse:
      type: object
      required: [job]
      properties:
        job:
          $ref: "#/components/schemas/ProcessingJobSummary"
        duplicate:
          type: boolean
      additionalProperties: false
...
    ProcessingDisplayStatus:
      type: string
      enum: [processing, error, ready]
    ProcessingStatus:
      type: object
      required:
        - claimId
        - status
        - claimNumber
        - insuredName
        - carrier
        - dateOfLoss
        - systemStatus
        - aiStatus
        - humanReviewStatus
        - job
      properties:
        claimId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/ProcessingDisplayStatus"
        error:
          type: string
          nullable: true
        claimNumber:
          type: string
        insuredName:
          type: string
        carrier:
          type: string
        dateOfLoss:
          type: string
        systemStatus:
          $ref: "#/components/schemas/SystemWorkflowStatus"
        aiStatus:
          $ref: "#/components/schemas/AiWorkflowStatus"
        humanReviewStatus:
          $ref: "#/components/schemas/HumanReviewStatus"
        job:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/ProcessingJobSummary"
      additionalProperties: false
    IngestRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
        carrier:
          type: string
      additionalProperties: false
    IngestClaim:
      type: object
      required: [id, claimNumber, status]
      properties:
        id:
          type: string
          format: uuid
        claimNumber:
          type: string
        status:
          $ref: "#/components/schemas/ClaimStatus"
      additionalProperties: false
    IngestDocument:
      type: object
      required: [id, fileName]
      properties:
        id:
          type: string
          format: uuid
        fileName:
          type: string
      additionalProperties: false
    IngestResponse:
      type: object
      required: [job]
      properties:
        claim:
          $ref: "#/components/schemas/IngestClaim"
        document:
          $ref: "#/components/schemas/IngestDocument"
        job:
          $ref: "#/components/schemas/ProcessingJobSummary"
        duplicate:
          type: boolean
      additionalProperties: false
    ReprocessClaimRequest:
      type: object
      required: [carrier]
      properties:
        carrier:
          type: string
          minLength: 1
      additionalProperties: false
...